pyproject.toml
[tool.poetry]
name = "graph based recomendation system"
version = "0.1.0"
description = "дипломный проект курса backend-разработчик Python Skypro"
authors = ["UltimaKIND <minion9@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
drf-yasg = "^1.21.8"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.10"
networkx = "^3.4.2"
numpy = "^2.1.3"
matplotlib = "^3.9.2"
mpld3 = "^0.5.10"
scipy = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
